backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai-whisper
torch
sentence-transformers
faiss-cpu
anthropic
elevenlabs
numpy
pdfplumber
ffmpeg-python
